+++ frontend/src/components/FeedbackForm/FeedbackForm.tsx
@@ -1,60 +1,35 @@
-import { ONE_SECOND } from '@constants/time.constant';
-import { EditableFeedbackType } from '@customType/feedback';
-import { currentVideoTimeState } from '@store/currentVideoTime.store';
-import { feedbackIdsState, feedbackIdxMapState, feedbackState } from '@store/feedback.store';
-import { mmssFormatter } from '@utils/common.util';
 import React, { useState } from 'react';
-import { useRecoilTransaction_UNSTABLE, useRecoilValue } from 'recoil';
+import { useRecoilValue } from 'recoil';
+
+import useAddFeedback from '@hooks/useAddFeedback';
+import { currentVideoTimeState } from '@store/currentVideoTime.store';
+
+import { ONE_SECOND } from '@constants/time.constant';
+import { mmssFormatter } from '@utils/common.util';
 
 import { fbFormWrapperStyle, fbInputStyle, fbStartTimeStyle } from './FeedbackForm.style';
 
 const FeedbackForm = () => {
 	const [inputVal, setInputVal] = useState('');
 	const [startTime, setStartTime] = useState(0);
-	const feedbackIdxMap = useRecoilValue(feedbackIdxMapState);
+
 	const currentVideoTime = useRecoilValue(currentVideoTimeState);
+	const { handleAddFeedback } = useAddFeedback();
+	const ENTER_KEY_CODE = 13;
 
 	const handleChange = (e) => {
-		if (!inputVal.length) {
-			setStartTime(currentVideoTime);
-		}
+		if (!inputVal.length) setStartTime(currentVideoTime);
 		setInputVal(e.target.value);
 	};
 
-	const handleAddFeedback = () => {
-		feedbackIdxMap.set(startTime, feedbackIdxMap.get(startTime) + 1 || 1);
-		const nextInnerIdx = feedbackIdxMap.get(startTime);
-		const newFeedbackId = generateFeedbackId(startTime, nextInnerIdx);
-		const newFeedback = {
-			id: newFeedbackId,
-			startTime: startTime,
-			innerIndex: nextInnerIdx,
-			content: inputVal,
-			readOnly: true,
-		};
-
-		insertFeedback(newFeedback, newFeedbackId);
-		setInputVal('');
-	};
-
-	const generateFeedbackId = (startTime: number, innerIdx: number) => {
-		return ('000000' + startTime).slice(-6) + ('00' + innerIdx).slice(-2);
-	};
-
-	const insertFeedback = useRecoilTransaction_UNSTABLE(
-		({ set }) =>
-			(newFeedback: EditableFeedbackType, newFeedbackId: string) => {
-				set(feedbackIdsState, (todoId) => todoId.concat(newFeedbackId).sort());
-				set(feedbackState(newFeedbackId), newFeedback);
-			},
-		[]
-	);
-
 	const handleKeyDown = (e) => {
-		if (e.keyCode === 13 && !e.shiftKey) {
-			e.preventDefault();
-			handleAddFeedback();
-		}
+		if (e.keyCode === ENTER_KEY_CODE && !e.shiftKey) addFeedback(e);
+	};
+
+	const addFeedback = (e) => {
+		e.preventDefault();
+		handleAddFeedback({ startTime, inputVal });
+		setInputVal('');
 	};
 
 	return (
